Extract Assets route module
