+++ frontend/src/components/assets/index.ts
@@ -1,0 +1,1 @@
+export { AssetsRoute } from "./AssetsRoute"
